// packages/tunnel-cast/lib/models/enums/field-procedure-type.enum.ts
export enum FieldProcedureType {
    ConditionalHandling = "CONDITIONAL_HANDLING",
    // ExistentsConstraint = 'EXISTENTS_CONSTRAINT',
    DefaultAssignment = "DEFAULT_ASSIGNMENT",
    Constraint = "CONSTRAINT",
    CustomConstraint = "CUSTOM_CONSTRAINT",
    Parser = "PARSER",
    CustomParser = "CUSTOM_PARSER",
    Transformer = "TRANSFORMER",
    CustomTransformer = "CUSTOM_TRANSFORMER",
}